+++ apps/server-web/src/locales/i18n/bg/translation.missing.json
@@ -11,5 +11,6 @@
   "Настройка": "Настройка",
   "относно": "относно",
   "Откажете се": "Откажете се",
-  "Status: Stopped": "Status: Stopped"
+  "Status: Stopped": "Status: Stopped",
+  "Ever Teams Server Web": "Ever Teams Server Web"
 }
