+++ apps/server-web/src/locales/i18n/en/translation.missing.json
@@ -15,5 +15,6 @@
   "Stop": "Stop",
   "Setting": "Setting",
   "About": "About",
-  "Quit": "Quit"
+  "Quit": "Quit",
+  "Ever Teams Server Web": "Ever Teams Server Web"
 }
