+++ apps/server-web/src/renderer/components/index.ts
@@ -1,6 +1,1 @@
-export * from './About';
 export * from './Popup';
-export * from './Server';
-export * from './SideBar';
-export * from './Updater';
-export * from './General';
